Refactor NNUE module structure
Bench: 1751379
